Basic styling for welcome page, removed unneeded images in public folder, renamed indexjs in capatcha folder to captchajs
