+++ routes/apiRoutes.js
@@ -19,9 +19,7 @@ router.get("/memes", (req,res) => {
 router.post("/insert",(req,res) => {
     db.MemeTemp.create({
         title: 'Hello',
-
         image: req.body.image
-
     })
     .then(console.log())
     .catch(err => res.status(422).end())
