+++ src/App.css
@@ -120,3 +120,32 @@ Button::after,
 	justify-content: space-around;
 
 }
+
+
+.fixed-top {
+	position: fixed;
+	top: 0;
+	right: 0;
+	left: 0;
+	z-index: 1030;	
+}
+
+.Route{
+	background-color: #222222
+}
+
+.footer {
+	position: fixed;
+	left: 0;
+	bottom: 0;
+	width: 100%;
+	background-color: #e9ecef;
+}
+
+.image {
+	float: left;
+}
+
+h1, h2{
+	text-align: center;
+}
