+++ src/components/pages/Welcome.js
@@ -1,25 +1,44 @@
 import React from "react";
 import { useHistory } from "react-router-dom";
-import Captcah from "../Captcha";
-import { Typography,Button } from 'antd';
-import Design from '../Design';
+import Captcah from "../Captcha/Captcha";
+
 
 function Welcome()  {
     const history = useHistory();
-    const { Title } = Typography;
 
     const handleClick = event => {
         event.preventDefault();
-        history.push("/MainPage");
+        history.push("/testpage2");
     }
 
     return (
-        <div className="Wcont">
-            <Design />
-            <Title class="h2" level={2}>Meme myself and I app!</Title>
-            <Title class="h3" level={3}>Create <span>Download</span> <span className="yellow">Share!</span></Title>
-            <Button className="GetStartedb" type="primary" onClick={handleClick}>Get Started</Button>
-            <Captcah />
+        <div className='container'>
+            <nav className="navbar navbar-expand-lg navbar-light bg-light fixed-top">
+                <div>Meme Myself and I</div>
+            </nav>
+            <div className='row d-flex justify-content-center'>
+                    <div className='jumbotron'>
+                        <h1>Meme, Myself, and I!</h1>
+                        <h2>Create, Download, Share</h2>
+                        <div className='d-flex justify-content-center'>
+                            <div className='col-md-6'>
+           ``                     <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Etiam non tellus lacinia, ultricies magna non, fringilla justo. Nullam ut consequat lacus, a aliquet ex. Cras a nisl finibus, commodo eros at, dictum sapien. Maecenas a elit fringilla mauris pellentesque ornare. Sed quis ligula pellentesque metus euismod lacinia ut vitae lectus. Cras blandit sed risus tristique tempor. Suspendisse bibendum nulla et dapibus pretium. Phasellus fermentum gravida tortor, ut posuere quam ullamcorper a. Integer quis enim lacus. Nulla pharetra ligula at lorem ultricies porttitor.</p>
+                            </div>
+                            <div className='col-md-6'>
+                                <img src='https://i.redd.it/qblm7745ex921.jpg'></img>
+                            </div>
+                        </div>
+                        <div className='d-flex justify-content-center'>
+                        <h3>Verify and start memeing</h3>
+                        </div>    
+                        <div className='d-flex justify-content-center'>
+                            <Captcah />
+                        </div>
+
+                    </div>
+            </div>
+                {/* <button className="GetStartedb btn btn-primary" type="primary" onClick={handleClick}>Get Started</button> */}
+            <div className="footer">footer</div>
         </div>
     );
